[IMP] hr_contract_salary: allow conditional advantages
Purpose: There are some business cases when one would like to make some
advantages dependent on others. Thus only allow to pick specific
(dependent) advantage if the mandatory advantage(s) are picked.
For example, the company might want to forbit to pick fuel card, in case
car is not picked.

To solve this, the m2m advantage_ids field is added on the
'hr.contract.salary.advantage' model.
All the mandatory advantages should be selected in order to select the
dependent advantage.

task - 2941827

closes odoo/enterprise#37713

Signed-off-by: Kevin Baptiste <kba@odoo.com>