[IMP] stock_barcode: get tracked product quants
When a product is scanned while doing an inventory adjustment, we search
if there is already exist a quant for this product at the given location
and if it's the case, the added line has the quant's on hand quantity.
That said, if the user scans a tracked product, it will look for a quant
for this product with no lot/serial number. As the quants for tracked
products usually have a LN/SN, no quant won't be found.
Therefore, it could be very useful to know the quantity for tracked
products too, and with this commit, it's what we'll do now:
- Search for product's quants (domain by product and location);
- Add a barcode line for each found quant who isn't already in the
  inventory adjustment.

task-3036810

Part-of: odoo/enterprise#33179