[FIX] account_invoice_extract: update for pdfjs compatibility
Previously, we were listening to the `pagerendered` event of pdfjs
to know when the box layers needed to be discarded.
In the more recent version of pdfjs, this event doesn't exist
anymore.

In this new version, the box layers will always be re-rendered, so
that we won't need this event anymore.  The cost of re-rendering it is
pretty much negligible and it keeps the code simple.

As it turns out, the owl conversion made hooking onto `pagerendered`
unnecessary: the previous version of the code used an inner
lazily-rendered element as its hook / base to attach the overlay
boxes, the owl version hooks onto `<div class="page">` elements, and
those are eagerly created by pdfjs (one per page), it's only filling
them which is lazy.

Since the boxes only need to be layered at the right location over the
page content, the actual rendering of the page content doesn't matter.

closes odoo/enterprise#31692

Related: odoo/odoo#100067
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>