[FIX] web_studio: don't alter the original view object from studio.
Edit a form view in studio, display the xml editor, go back to the normal
webClient

Before this commit, the Model of the formView was potentially altered programmatically
when switching the edition mode of studio.

After this commit, there is no risk, as nothing is ever written onto the view object from
the view registry.

closes odoo/enterprise#33033

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>