[IMP] helpdesk: remove customization for rating.rating view
rating.rating form view: The customization used in helpdesk is removed
to uniformize this form with other apps. The cosmetic change in the
form should be done directly in rating.rating. The result here is that
the comment field will only takes half the window in the helpdesk
module (as it was the case in the project module).

task-3034806
related PR: https://github.com/odoo/odoo/pull/104250

closes odoo/enterprise#33309

Related: odoo/odoo#104250
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>