[FIX] web_mobile: saving user profile hangs on iOS app
Due to our inability to update the iOS app (following review from
Apple), this commit aims at workingaround this limitation by avoiding to
wait for a Promise that is never resolved in the iOS app.

Steps to reproduce:
- Setup Odoo with hr & documents installed
- Open it with the iOS mobile app
- In the BurgerMenu, open "My Profile"
- In the stat buttons, click/tap on "x Documents"
=> the whole view hangs

Note: this commit also replicates a comment from legacy codebase
explaining why the Promise returned by `downloadFile()` is voluntarily
not awaited.

closes odoo/enterprise#33133

X-original-commit: 73149c2920237365f1d39b5f1267e499c5e22131
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans (app) <app@odoo.com>