[FIX] timesheet_grid: display pause button when timer is running
Steps :
Go to Timsheet.
In grid view, press play button and add a project.
Go to list view and check the newly created line.

Issue :
You can't pause the timer.

Cause :
On top of the conditions that display_timer is true
and the record is not selected,
another condition means to check that the field is not readonly,
in which case it would make sens to not display the button.
Yet, on one hand, this widget is never used for readonly fields
and, on the other hand, record.isReadonly(fieldName)
returns true when the timer is running.

Fix :
Rework these conditions:
display_timer still has to be true
the line isn't selected and/or the timer is running.

task-3037181

closes odoo/enterprise#33058

X-original-commit: 740d28b1a25b884912eed66a23c422a359831e95
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Audric Onockx (auon) <auon@odoo.com>