[IMP] web,*: force props validation for components
*: mrp_workorder, spreadsheet_edition, voip, web_enterprise, web_gantt,
    web_studio

This commit adds a warning if the props validation is not set for a
component.

The props validation is important to tell how a component should be
used, by looking at its code, it's a good documentation of the component.
It's also critical, to test if the component is correctly used, if all
the obligatory props are passed and that there are of the correct type.

For more information, see: https://github.com/odoo/owl/blob/master/doc/reference/props.md#props-validation

closes odoo/enterprise#33044

Related: odoo/odoo#103723
Signed-off-by: Géry Debongnie <ged@odoo.com>