[FIX] stock_barcode: select right quant by loc
How to reproduce:
- Create a product tracked by lots;
- Add quantities for this product with the same lot in two locations;
- Mark them as needing a recount;
- Open the Inventory Adjustment in the Barcode App -> You should get two
  lines for this product, each in a different location;
- Set the right quantity for one on them, then quit the Inv. Adjustment
  and go back in it.
- Scan the location of the line with the set quantity -> It should set
  the scanned location in bold on the line.
- Scan the product's barcode -> It select the product in the other
  location instead of select the one in the scanned location.

This issue was due that when it looks to select a line, it ignores the
line in the right location because this line is completed (which is
wanted) and then, it finds the line in another location and select it
because the line has no counted quantity yet.
This behavior is wanted for the picking (because the source location of
the selected line will be overrided by the previous scanned one) but can
be problematic for inventory adjustment if the user has to count the
same product in multiple locations.

Part-of: odoo/enterprise#33084