[IMP] hr_contract_salary: clean contract PDF template
The PDF template (employee_contract.pdf) has a default signature for the HR responsible that is
part of the PDF file.
When a contract is signed, the signature from the Sign module overlays with the one from the PDF.

This commit changes the PDF emplyee_contract.pdf to remove the signature in the PDF file.

task-3037464

closes odoo/enterprise#33202

Signed-off-by: Kevin Baptiste <kba@odoo.com>