[IMP] planning: add switch shift functionality
This commit adds a "switch shift" functionality to the
planning app.

This new functionality allows the employees of a company
to ask for a switch change by clicking a button on the
shift form view either in the front end or in the back end.

Asking for a switch will make the shift available for other
users to assign themselves to. If someone does, the original
assigned employee will be notified through an automatic
email.

Task-3035989

closes odoo/enterprise#34565

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>