[FIX] stock_barcode: unfold dont select other line
When the user unfolds a group of lines, the parent line is automatically
selected, which is equal to select one of the subline.
The issue was, when doing that, sometime a subline of the group is
already selected and by unfolding it, it's another subline who will be
selected for no reason.

How to reproduce:
- Create a product tracked by lot;
- Add some quantity on hand for this product with two lots. E.g.:
  - lot1 x2
  - lot2 x2
- Create and confirm a delivery for this product, only for 2 qty (so,
  only lot1 will be reserved);
- Open this delivery in the barcode app;
- Scan 1 time the lot1;
- Scan lot2 -> A second line is added for the tracked product for this
  lot and this line is selected, but the lot1 and lot2 lines are grouped
  so for now, we only see the parent line as the selected one.
- Click on the button to unfold the line and display the sublines
  -> The line for lot1 is now selected. As the lot2 was the last scanned
     thing and the lot1 line was not clicked by the user, the line for
     the lot2 should still be the selected one.

runbot error id: 8274

Part-of: odoo/enterprise#34104