[ADD] l10n_cl_edi: l10n_cl_edi_stock, l10n_cl_edi_boletas: new demo mode
Improve testing workflows for customers, demo, and training purposes

Customers want to test with their data and duplicated DB but real certificates are sensible to be used when no validation for the SII is needed

Why not use demo certificates?: there are no demo certificates. All are real certificates with sensitive personal data

Also in the Chilean localization, there is an extra complication since CAFs for each type of document are needed in testing and usually, they are limited by the government

Proposed solution:

We have an extra environment called demo, where no restrictions regarding certificates or CAFs are present to be able to test and create demos of every workflow with the creation of the XML. This environment will not return any SII errors. It is only for local testing and demo/training purposes.

closes odoo/enterprise#36098

X-original-commit: 7af01734505667c444d547c12fb12bf40b791228
Signed-off-by: Josse Colpaert <jco@odoo.com>