[FIX] spreadsheet_edition: prevent the selection of an offset without a field
Currently, a user could create a date filter without selecting a field
for a datasource but still select an offset. Notwithstanding the
traceback that ensues, this does not make sense from a a functional POV.

This commit deactivates the offset selection while a field has not been
selected by the user.

How to reproduce:

From a sheet with a pivot:
- create a date filter
- do not set field
- set offset to any (non null) value
- Save the filter
=> trackback

Part of 3002282

closes odoo/enterprise#34799

X-original-commit: 0412186a675960a8f4a83398cf95c628203aa3fa
Related: odoo/odoo#107485
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>