[FIX] web_studio: set nodes invisible instead of removing in studio
Current behaviour:
Editing names of form tabs in studio doesn't apply the changes
to the correct tab. Sometimes you can't even add fields to a form
sheet, or it adds to the incorrect ones. Sometimes you get
stackstraces from xpath that couldn't be found.

Expected behaviour:
Adding/Editing elements in studio shouldn't produce any errors,
or change other elements in proximity.

Steps to reproduce:
- Install Inventory and Studio
- Disable Variants in Settings
- Select a Product and go in Studio
- Add a new Tab, try editing it's name
- Observe that not the New Tab name changed, but the one of it's left.

Reason for the problem:
There is incoherency between the view that Studio is getting from the server
(nodes to which the user doesn't have access are not present in the view),
and the view that the server is applying Studio's changes.
(there is no permission checking on the view used to apply the patches onto).
This leads to WOWL generating incorrect xpaths of elements, because it is basing
itself on the view sent from the server (with some nodes missing), and then the
server is trying to apply the changes with that xpath on a view that is not the same.

Fix:
In the context of studio, instead of removing the nodes we don't have access to,
we set them invisible. This is more inline with behaviour in saas-15.3 and before.

Affected versions:
- 16.0
- master

opw-3018490

closes odoo/enterprise#33364

X-original-commit: f3f620c5fbd6ad11169ebd0ebeecd3afda6af397
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>