[FIX] {snailmail_}account_followup: fix overdue invoices and attachments, and improve views layout
There is an issue with the followup status, where an invoice is set as overdue when the due date is today. It shouldn't be overdue in such a case.
When attaching invoices to a followup letter (manually, via the wizard), every invoice is attached. We should only attach invoices that are actually selected.
The same goes for automatic reminders, only invoices that have already been sent/printed should be attached.

This commit fixes these 3 issues and also improves the layout of the followup wizard and the followup level views view (mainly reordering fields, to better match the fix about attached invoices).

closes odoo/enterprise#33708

X-original-commit: 78cf02b9bbebc0fe6106ff6ff0552879d72a1b54
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>