[FIX] knowledge_website: fix home route with no articles in frontend
Bug description:
----------------
When accessing `/knowledge/home` with a portal user when the user has not
access to any article, one gets an error 500.

Cause:
------
knowledge_website frontend template overrides a rendering condition of the
overridden template, causing the favorite button to be rendered even if no
article is given. The id of the article is used to render this button, which
fails when no article is given.

Fix:
----
Add the rendering condition of the overridden template in the
knowledge_website template as well.

Task-3040876

closes odoo/enterprise#33213

X-original-commit: 72544496202a9b96597c7631781c40e542154ace
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>