[FIX] industry_fsm_sale: fix failed test during at_install
Before this commit, when the tests are launched during the first
installation of modules. A test launched during the installation of the
`industry_fsm_sale` failed. The reason is because the `account` has no
journal and so it is impossible to create some invoices.

This commit changes the moment in which the test is launched. The test
will be launched after the installation of all modules and so the
`l10n_generic_coa` module will be installed and a journal will be there
by default to be able to create invoices.

closes odoo/enterprise#34942

X-original-commit: ba0779d81809a6a8e7eee85cc042b51b2778e30c
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>