[FIX] documents_account: documents app sending customer information
Steps to reproduce:

- We go to documents app > Configuration > Settings > Accounting, and
activate this option with workspace as Finance.

- Then we go to journal (just under the previous option) and we create
a journal for Customer invoices within Finance workspace.

- We go back to Documents > Finance > Mark Inbox in tags and upload a
new file.

- We select a contact for this file and we select Create Customer
Invoice.

Issue:

After all the steps we're gonna see that the contact that we set is not
set as the customer of the invoice created, it get lost.

Solution:

The easies solution I found so far is to change the order we get the
partner_id in the create_record method. This is because after
_create_document_from_attachment the document.partner_id is overridden
and we lost the one that we had initially.

opw-2991015

closes odoo/enterprise#33672

X-original-commit: bfdc6eb525c38cad8925d1ce9b24ceaee824617b
Signed-off-by: Grazioso Andrea (agr) <agr@odoo.com>
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>