[FIX] web_studio: a kanban editor for the new WOWL KanbanView
Before this commit, the kanban editor was still in legacy, which was not ideal since some widgets
and field widgets available in the new implementation were not displayed in the kanban editor.
This produced misleading output in the editor, or eventually crashes in the worst cases.

After this commit, the kanban editor is fully written in the new implementation, on the same premises
as the formEditor[1], that is, the xpath of each node of interest in the arch is computed and used
to place hooks, identify a node and edit it and generally to edit the view.

[1]: 370ba732e84b8d121e3283a16dbf233d0c7ac4c8

closes odoo/enterprise#33181

Related: odoo/odoo#104116
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>