[REM] account_predictive_bills: merge into account_accountant
The prediction is now merged into the account_accountant module.
The account prediction boolean setting (which was used nowhere)
is now replaced by a boolean for product prediction.

The context keyword `account_predictive_bills_disable_prediction` is
now removed.

The account predictive feature is activated by default on bills, and
activated on invoice if it is quick-editable.

closes odoo/enterprise#31640

Related: odoo/odoo#104625
Related: odoo/upgrade#3918
Signed-off-by: William André (wan) <wan@odoo.com>