[FIX] project_forecast: restrict project update button to planned slots
This commit purpose is make the behavior of all planning views coherent.
The 'planned shifts' button form the project update redirect to a
planning view where the shifts with no planned date are visible in the
list view. It is expected not to, as it is for all the other planning
views.

This commit :
 - change the domain of the action so that the shifts with no planned
date are not picked

task 2992614

closes odoo/enterprise#33356

X-original-commit: 47b38740705cbfa716975e3140d661bb120c5684
Related: odoo/odoo#104386
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>