[FIX] l10n_mx_edi_40: do not put taxes when 100% discount
Before, when having a 100% discount, it would put it as ObjetoImp 02.
But then it would complain because for 02, it expects a child node,
with the Impuestos while there are none in the tax_details.

The solution is to put ObjetoImp 01 instead, so it does not expect
the Impuestos anymore when the tag is zero.

opw-3001936

closes odoo/enterprise#34858

X-original-commit: efe9a1f7be1963be11ebfdf18ebcb346e497c4ae
Signed-off-by: John Laterre (jol) <jol@odoo.com>