[FIX] knowledge: add safecheck before reading wysiwyg instance
In the behaviors of Knowledge, we retrieve the root node of the article
content by using the wysiwyg instance given to the props of the behavior
component. However, this object is not instantiated when the article is
opened in readonly mode. As a result, an error can be raised if the
behavior accesses a property of the wysiwyg instance in readonly mode.

To fix the issue, we will add some safe checks in the existing behaviors
and we will also pass the root HTML node of the article content to the
props of the behavior components. The component can then safely use that
reference to do any computation involving the article content (e.g:
fetching the heading of the article, etc).

Step to reproduce the error:
- Open an article
- Add a table of content with the /toc command
- Lock the article
- Reload the page

=> An error is raised indicating that the wysiwyg editor instance is undefined.

task-3044528

closes odoo/enterprise#33264

X-original-commit: 6594403dbe77936165e9707537355924f4817597
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>