[FIX] account_sepa_direct_debit: make PmtInfId unique across messages
Even if the official regulation is stating that PmtInfId must be unique
for each payment within the message, it seems that some banks are
refusing some files because the same PmtInfId is used across several
messages. We should make it unique by using MsgId inside it to avoid
this situation. This is already the case in credit transfer files, so
this commit allows to align the situation between the two schemes.

closes odoo/enterprise#33354

X-original-commit: 09d6b4f39f23e4b682caae1c7b414f7182bf72dc
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>