[FIX] knowledge: add custom switchView method for embedded views
When an embedded view calls the `switchView` method of the JS framework,
the system will try to switch the view of the main view rather than the
view shown in the embedded view. With that behavior, an error can be raised
if the main view does not support the given view type.

To fix the issue, we will overwrite the `switchView` method of the embedded
views to execute a new action window to open the view we want rather
than trying to "switch" the current view. This will ensure that the user
be able to open the view even if the embedded view does not support the
view type that needs to be loaded.

Step to reproduce the error:
1. Insert a map view containing pins in a Knowledge article.
2. Open the article in which you have inserted the map.
3. Click on a pin linked to a single record.
4. Click on the "open" button of the open tooltip.
=> An error is thrown as the Knowledge editor does not have any form view.

task-3012552

Part-of: odoo/enterprise#32627