[FIX] sale_ebay: update ebay server when delivering
When validating a picking, if it is related to a SO generated from Ebay,
we send a message to their server to update the shipment state as done.
However, this update should only be done when validating the picking to
the customer. Otherwise, suppose a 2-step delivery configuration: on
Ebay side, the shipment will be considered as done as soon as the pick
step is done, which is incorrect

OPW-3011272

closes odoo/enterprise#33341

X-original-commit: 39580228510eab5cb7f1b5ab00310b641549be68
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>