[FIX] hr_*_extract: digitization when auto_send
Before this commit, the recruitment_extract_show_ocr_option_selection
value was not verified in the function _message_set_main_attachment_id.

This is needed because `message_post` from the parent `mail.thread`,
calls this function, and we do not want to send the attachment upon post
to the OCR if the recruitment_extract_show_ocr_option_selection is not
`auto_send`.

Also, the hr_expense_extract module was updated to match the changes
brought in hr_recruitment_extract.

X-original-commit: 01f0e773ad866b1629ea433d08fe9d7af25bc88b
Part-of: odoo/enterprise#34252