[FIX] timesheet_grid: ensure toggle button works as expected
Prior to this commit, the visibility of the toggle button was managed
by a `t-if` that was bound to the state of the component.
This was causing the mouseout event not to be triggered when moving
the mouse fastly. Indeed as the mouseover changes the component state,
the components gets re-rendered and if the process is triggered at the
time when the mouse is leaving the component, the mouseout is not triggered.

There is no other solution using Owl only, that's why we have chosen
to use css.

Unfortunately the test can't be adapted so that the css is tested so
we had to drop it.

closes odoo/enterprise#33325

X-original-commit: b8a0c9f939d73938596dc474c5da802b474e7579
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>