[FIX] sale_subscription: invoice only contract in sale state
Before this commit, contract in sale and done state could be invoiced if they
filled the conditions (end_date in the future, next invoice date reached,
...). Such case did not happened in normal situation for the locked SO.
But if a bug happened and resetted the end_date of a locked order, it would
reach its next_invoice date and would be invoiced.

To prevent this issue, we make sure to only invoice the contracts in sale state.
For now, we don't have any reason to invoice locked orders as they are invoiced
in advance before their renewal are confirmed.

We can't modify the _recurring_invoice_domain method to only fetch contract
in sale state because done order needs to be searched and closed automatically
when they reach their end_date.

closes odoo/enterprise#33825

X-original-commit: f2e2220c1fc5dacd088370a1ffb8daf80b6e3a26
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Arnaud Joset <arj@odoo.com>