[FIX] helpdesk: restric rating domain
This commit purpose is to hide rating for the user when the user's
active companies are different than the company of the rating's
associated employee. This fix only works when the user go to the page
through the menu item. Due to technical limits, if the user change the
selected companies while already on the view, the domain that will apply
is the domain of the view, which can not be improved.

This commit :
 - change the domain of the view so that the rating whose team's company
are different from the user's selected companies are not shown

task 3044723

closes odoo/enterprise#34775

X-original-commit: 69be8a8261c9b253459262d15b236881712bef7d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>