[FIX] mrp_workorder: Redesign "instructions" column in BoM
Since 1a6abc7, 'Instructions' column in the BoM 'Operations' tab is displayed as the first column and there's too much space between "Instructions" and the value of quality_point_count.
It also has an impact on the display of the 'Add a line' and 'Copy Existing Operations' at the bottom of the tree/list.
This PR place the 'Instructions' column as shown there : https://tinyurl.com/yedpuxwj.
It also revert to the previous design, replacing the 'Instructions' string placed before the value of quality_point_count by a list icon placed after.

closes odoo/enterprise#33625

Task: 2985735
X-original-commit: a465f08b1df5868438463413c52270bfc7cbd07b
Signed-off-by: Steve Van Essche <svs@odoo.com>