[FIX] appointment: avoid error on deleted onboarding steps
And trigger refresh of client's onboarding panel (see below).

In two situations, we didn't handle well the case where an opened onboarding
step is deleted after it was opened but before it was validated:

* 2nd onboarding step (link): The returned value of
 `AppointmentOnboardingLink.search_or_create_onboarding_invite` should include
 the (valid) invite url created.

* 3rd onboarding step (provider config): Using `raise_if_not_found=False` is
not sufficient, we also mustn't use the result if it evaluates to None.

This also revealed the importance of returning wasFirstValidation True in these
cases in order to force a refresh of the onboarding panel.

 Task-3046707

closes odoo/enterprise#34073

X-original-commit: 36096c65c31a431db91e123a1c1a3f719175207c
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>