[IMP] sale_subscription: use sale.order.alert to manage contract health
Before this commit, the subscription health domain were defined on the
subscription templates. Health was computed during the KPI compulation.

This approach has some drawbacks:
- It prevents to have health values on contracts without quotation templates
- It is not logical to defined the health conditions there.

As the sale.order.alert system is powerful enough to handle the health settings,
we decided to use it to remove fields from the domains and to promote the
alert system which could be more used.

Taskid: 2831415
Part-of: odoo/enterprise#33239