[IMP] sale_subscription: improve the 'to_renew' conditions with alerts
Before this commit, the cron cron_subscription_expiration was managing
the contracts to set as "to renew" one month before their end date.

It was not configurable and not clear when this flag was triggered.

This commit replace the logic of the cron by creating a dedicated
sale.order.alert that could be understood and modified by the customer
more easily.

Taskid: 2831415
Part-of: odoo/enterprise#33239