[IMP] documents_project: link uploaded documents to projects
When uploading a document through the portal or the documents app,
we automatically link them to the relevant project or task.

However, we don't do anything when uploading a document through
a share link.

With this commit, documents uploaded through a share link will
be automtically linked to a relevant project if there is any.

Task-3045852

closes odoo/enterprise#33367

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>