[FIX] helpdesk,hr_payroll,...: Avoid using write on compute methods
Purpose
=======

Onchange methods are executed in "edit-mode", so you need to save to get the definitive
values saved on the DB, but if you do a write on it, that changes are immediately saved
to the DB, so if you discard the changes on UI, you will have inconsistent data. That's
why you have to do update instead, that only modifies the temporary dataset.

Courtesy of https://github.com/OCA/pylint-odoo/issues/356

closes odoo/enterprise#33366

Taskid: 3046426
Related: odoo/odoo#104419
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>