[FIX] account_bank_statement_import_camt: unique import id across files
Currently, when importing CAMT files, we are building a unique import id
for each statement line by using the elements AcctSvcrRef and NtryRef.
However, it could happen that these elements are the same for another
statement line on the same IBAN account/journal. We should include the
statement name in the unique import id to correctly identify these two
distinct operations instead of preventing the import of the operations,
forcing the user to manually change the file.

closes odoo/enterprise#33470

X-original-commit: cad02743089ad88576f3cddbc40184f6ad43ba09
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>