[FIX] account_bank_statement_import_camt: unique import id across files
Currently, when importing CAMT files, we are building a unique import id
for each statement line by using the elements AcctSvcrRef and NtryRef.
However, it could happen that these elements are the same for another
statement line on the same IBAN account/journal. We should include the
statement name in the unique import id to correctly identify these two
distinct operations instead of preventing the import of the operations,
forcing the user to manually change the file.

closes odoo/enterprise#33475

X-original-commit: 164030ba67f51404c48d565e6ab14c5a2335bc57
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>