[FIX] account_reports: cash flow statement
Before this commit, the cash flow report was only using account moves
that were done on the bank and cash journals. That behavior was not
correct because the cash flow report should use all moves made with
bank and cash accounts, whatever the journal.

So if an account move is done in a general journal (not bank or cash)
but uses a bank or cash account, it should be use on the report.

closes odoo/enterprise#33402

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>