[FIX] quality_mrp_workorder: Allow pass and Auto validate failed QCP in tablet view
**Steps to reproduce :**
- Create a BoM with an operation containing a passfail Quality Check.
- Create an MO with the previously created BoM and confirm it.
- Enter in the tablet view of the just created Work Order.
- Set the quality check as 'Fail'.
- Change the same quality check to set it as 'Pass'.

**Expected Behavior**
- When clicking 'Fail' : The 'Fail' is directly confirmed and the next step is loaded.
- When clicking 'Pass' : The `quality_state` should change from `'fail'` to `'pass'`.

**Actual behavior :**
When clicking 'Fail' :
- The QCP page update and a new button 'Next' appears.
- The user can confirm that the quality check is failed by clicking on the 'Next' button, or he can press 'Pass' to correct a misinput.
- He can eventually press again 'Fail' though it's useless.
- Either he press 'Next' or 'Pass', the next step is loaded.

When clicking 'Pass' :
The `quality_state` value stay `'fail'`.

https://github.com/odoo/enterprise/blob/316238b28539575d4917265e8e637d2dcda84d32/quality_mrp_workorder/views/quality_views.xml#L97-L98
Unlike the 'Fail' button that calls `do_fail()`, the 'Pass' button calls `action_next()` instead of `do_pass()`.
`action_next()` calls `_next()` that set the value of `quality_state` to `'pass'`, but only if `quality_state` wasn't already defined (an so equal to `'none'`).
https://github.com/odoo/enterprise/blob/316238b28539575d4917265e8e637d2dcda84d32/mrp_workorder/models/quality.py#L465-L466
This behavior is used by quality checks that have 2 possible states : `'none', 'pass'`, such as upload a picture, take measure... But is not intended to be used when there's more than 2 possible states, as it's the case in the `'passfail'` test_type.
We so make the 'Pass' button calls the newly introduced ´action_pass_and_next()`, that calls ´action_next()` as it updates

closes odoo/enterprise#34121

Task: 2985735
X-original-commit: b3461120b15dbff14600549ef466b26571b9881b
Signed-off-by: Steve Van Essche <svs@odoo.com>