[FIX] stock_barcode: scan an already-used package
If a package already has a location, the user will not be able to use it
as destination package in the Barcode app.

To reproduce the issue:
1. In Settings, enable "Packages"
2. Create a package PK
3. Create a product P
    - Storable
    - With a barcode
4. Update the on hand quantity of P:
    - 1 x P in package PK
5. Barcode > Operations > Receipts, Create
6. Scan P
7. Scan PK

Error: an error message is displayed "You are expected to scan more
products or a destination location."

If the user edits the line added at step 6, he will be able to select
the package PK. This flow also works from the Inventory app. Therefore,
scanning PK should also work.

Server-side, when loading the view state, we get the available packages
for each picking:
https://github.com/odoo/enterprise/blob/b2080c9aa9d8318d381d9e0f5a1b4a94eb86a258/stock_barcode/models/stock_picking.py#L69-L70
However, the domain is incorrect: it gets all packages that do not have
any location yet. It should also include all packages that are already
at the destination location of the related picking.

OPW-2997713

closes odoo/enterprise#34164

X-original-commit: fcb8ca35e1b78c471d7a78656f49f751c1eee9c6
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>