[FIX] l10n_au_aba: update domain of filter
Current behavior:
When we open payments to send via ABA from the accounting dashboard, there is a
the difference in the number of records from the list view and dashboard which
happens to be the same.

Expected behavior:
When we open payments to send via ABA from the accounting dashboard, the number
of records should be the same in the dashboard and list view.

Reason for that: We have changed the domain of the dashboard view but the list
view's domain is not updated.

The improvement in the domain is from this PR (#20177). But the same thing is
not reflected on the list view's domain.

Fix
To update the domain of the filter of the list view.

Task:- 3017513

closes odoo/enterprise#33518

X-original-commit: 7f5c966ba2a0aac3caa120bfeb76cbaeaa2afc36
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>