[FIX] account_invoice_extract: wrong currency
Before this commit, we were taking the first found currency that
resembles to what the OCR found.

This may cause an issue with currencies that have the same symbol or the
same name. For example  with the Danish Krone/Norwegian Krone or the
Australian dollars/American dollars.

With this commit, the currency is chosen by order of priority. We still
try to find the correct matches depending on what the OCR found, but
amongst those results, we first choose the one that was used in the last
invoice with the same customer, then the one that is linked to our
company, and finally the first item of the list if none of the former
ones were found.

Also, the usage of `ilike` does not seem appropriate in this currency
selector:
 - If the OCR returns "kr" as the currency, we could select by mistake
the Pakistani rupee as it has the "PKR" currency code.
However, to avoid breaking anything, or changing the current behavior,
after using `=ilike`, if no `possible_currency` was found, we still use
`ilike`.

task-3048098

closes odoo/enterprise#33823

X-original-commit: 1064d24ffc7dd6ecd7545ce816dfb8e0cf2ceb9f
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Signed-off-by: Arthur Gossuin (goa) <goa@odoo.com>