[FIX] l10n_co_edi: download signed invoices when checking status fast
Current behaviour:
Clicking too rapidly on "Process Now" when creating an invoice
can generate errors and block the invoice for further process
by Carvajal (Columbian government tax agency)

Expected behaviour:
Invoice shouldn't be blocked when clicking too quickly after
creation on "Process Now".

Steps to reproduce:
Check referenced ticket for setup

Reason for the problem:
Sometimes the `legalStatus` of the response is empty,
because Carvajal is still processing the legality of the invoice,
and we were flagging this response as a rejections.
When we interpret a response as a rejection, we retry to generate a new XML
for the invoice, but then Carvajal responds with a (real this time) error
saying an entry in their system for this company + invoice number is already present.

Fix:
Modify the conditions so they more explicitly handle the different cases between
an invoice that is accepted/in progress/rejected.

Affected versions:
- 15.0
- saas-15.2
- saas-15.3
- 16.0
- master

opw-2851945

closes odoo/enterprise#36924

X-original-commit: e8b84073b6c960428d4efc46630e4a1800e97ab2
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>