[FIX] l10n_mx_reports: fix the diot report exports test
In some cases, the order of the lines outputted by the file export
of the diot report could be different. This causes issue with the
test as it is expecting the same order each time.
This change will avoid this issue by re-making the text output
with the line being ordered.

closes odoo/enterprise#33836

X-original-commit: ea9db5bb9fe4e4a1ce62ae2b6c3ea47b3f1d2e2c
Related: odoo/odoo#105523
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>