[FIX] l10n_mx_reports: fix the diot report exports test
In some cases, the order of the lines outputted by the file export
of the diot report could be different. This causes issue with the
test as it is expecting the same order each time.
This change will avoid this issue by re-making the text output
with the line being ordered.

closes odoo/enterprise#33516

Related: odoo/odoo#104755
Signed-off-by: Laurent Smet <las@odoo.com>