[IMP] l10n_pe_edi, voip: remove useless upgrade_boolean widget
In a Community database, in Settings, when checking a config exclusive
to Enterprise, the Upgrade Boolean field is used to display a popup
suggesting the user to upgrade to Enterprise.

The presence of this field does not make sense for modules that are
Enterprise-only, such as VoIP.

This commit deletes occurrences of the Upgrade Boolean field in
Enterprise modules, as they are not relevant there.

closes odoo/enterprise#34120

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>