[FIX] timesheet_grid: read worked hours for employees in companies only
Steps :
Have 2 Companies (C1 & C2).
Create an Employee E2 in company C2.
Login with E2, both Companies checked.
Register a Timesheet in a Project from C1.
Log in with Admin, only C1 checked.
Go to All Timesheets.

Issue :
Multi-company access error.

Cause :
Because his Timesheet is in the view,
we try to access his worked hours to make a widget.
The access is denied because his Company is not in allowed companies.

Fix :
Make this widget only for Employees that are in allowed companies.

task-3053612

closes odoo/enterprise#33526

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>