[FIX] planning: hide the template_creation label if no planning manager
Before this commit, the following commit 4f98d3bb9e8
reviews the form view to keep the previous behavior before the OWL
conversion. However, this commit lets `template_creation` label visible
if the current user is not a Planning Manager, so the label will be
visible and not the field.

This commit hides the label is the current user is not a Planning
Manager.

closes odoo/enterprise#33580

X-original-commit: a917126454f1f45e208ce918cf4581335dc69008
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>