[FIX] planning: hide the template_creation label if no planning manager
Before this commit, the following commit 4f98d3bb9e8
reviews the form view to keep the previous behavior before the OWL
conversion. However, this commit lets `template_creation` label visible
if the current user is not a Planning Manager, so the label will be
visible and not the field.

This commit hides the label is the current user is not a Planning
Manager.

closes odoo/enterprise#33539

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>