[FIX] *: prevent default project_id on non timesheets
*: helpdesk_timesheet, timesheet_grid

Prior to this commit and since odoo/odoo#94144, a default project_id is
provided event for non timesheets.

This commit fixes this issue by setting a context key in the `account.analytic.line`
actions that are related to timesheets and by setting the default project_id only
when this key is set to `True`.

closes odoo/enterprise#33550

Related: odoo/odoo#104820
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>