[IMP] planning: change self assign button colour based on role
This commit allows the self-assign button (which appears in open
shifts) to change between primary and secondary depending whether
the slots role is covered by the current user's roles. Also we
changed its placement depending on whether it is primary or
secondary. Also we restrict the visibility of the button to non-
admin users.

Task-3037867

closes odoo/enterprise#33551

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>